Initial recording store class with unit tests

+++ android/mobile/src/androidTest/java/com/salidasoftware/gpsrecording/LineDAOTest.kt
@@ -68,7 +68,11 @@ class LineDAOTest {
         val beforeCount = lineDAO.count()
 
         val found = trackDAO.getById(trackId)
-        trackDAO.delete(found)
+        if (found != null) {
+            trackDAO.delete(found)
+        } else {
+            Assert.fail("Got an unexpected null track!")
+        }
 
         val afterCount = lineDAO.count()
